libstagefright: Avoid cache thrashing by preserving more cached data
Instead of a hard-coded 1MB "gray area" let's set a default of 4MB
and adapt it based on any dynamic cache parameters.

At 1MB there are many videos that have mutiple streams and the
streams appear in very wide byte ranges of the data.  For example,
episode 6 of http://kissanime.com/Anime/Dragon-Ball-Super (when viewed
at 720p) which has an almost 3MB gap between the audio and video
streams.

During playback we buffer the default 20MB of data and then after
playing that data we throw it all (except 1MB) away and then we
end up seeking because the data for the earlier stream is no longer
available.

Rather than just hard-coding the value, let's try to set some sensible
ranges based on the low/high water thresholds.

Change-Id: Ic8f01b04215552157fda2549eab99c41b8ff5116