Implement support for RTSP over IPv6
When trying to stream RTSP over IPv6 a CHECK was triggered
during the parsing of the IPv6 url, since the RTSP implementation
did not support IPv6.

This patch implements the missing functionality needed in order to
support RTSP streaming over IPv6.

Parsing of IPv6 url's is now added. Previously only IPv4 addresses
were supported.

The socket setup procedure is also modified. The gethostbyname() API's
are deprecated and are now replaced with the getaddrinfo() API,
which supports both IPv4 and IPv6 address resolution.

All possible resolutions will now be tried in the order provided by
getaddrinfo() until a connection is established.

Change-Id: Ie231bff2966bfa037659f4512d653e27c392a315